Re-organizing linker tests into smaller files and folders
