+++ tests/linker/variables/link.mixin.test.js
@@ -1,0 +1,17 @@
+import { expectScopeOf } from '../link-expects'
+import { MixinDeclaration } from '../../../src/model'
+
+describe('Class scoping', () => {
+
+  it('Mixin scope includes instance variables', () => {
+    expectScopeOf(`
+        mixin Bird {
+          const energy = 23
+        }
+      `,
+      MixinDeclaration, m => m.name === 'Bird',
+      ['energy']
+    )
+  })
+
+})
